pkg/synnergy_network/core/compliance/transaction_monitoring/transaction_monitoring.go
package transaction_monitoring

import (
	"context"
	"database/sql"
	"encoding/json"
	"log"
	"net/http"
	"sync"
	"time"

	"github.com/nats-io/nats.go"
	_ "github.com/lib/pq"
)

// Transaction represents a blockchain transaction
type Transaction struct {
	ID        string    `json:"id"`
	UserID    string    `json:"user_id"`
	Timestamp time.Time `json:"timestamp"`
	Amount    float64   `json:"amount"`
	Type      string    `json:"type"`
	Status    string    `json:"status"`
	Category  string    `json:"category"`
}

// Anomaly represents a detected anomaly in transactions
type Anomaly struct {
	TransactionID string    `json:"transaction_id"`
	DetectedAt    time.Time `json:"detected_at"`
	Reason        string    `json:"reason"`
}

// TransactionMonitoringSystem manages the transaction monitoring process
type TransactionMonitoringSystem struct {
	db            *sql.DB
	natsConn      *nats.Conn
	alertCh       chan Anomaly
	classifyFunc  func(Transaction) string
	model         *PredictiveModel
	dashboardData DashboardData
	dataLock      sync.Mutex
}

// NewTransactionMonitoringSystem initializes a new transaction monitoring system
func NewTransactionMonitoringSystem(db *sql.DB, natsURL string, classifyFunc func(Transaction) string) (*TransactionMonitoringSystem, error) {
	nc, err := nats.Connect(natsURL)
	if err != nil {
		return nil, err
	}
	return &TransactionMonitoringSystem{
		db:           db,
		natsConn:     nc,
		alertCh:      make(chan Anomaly, 100),
		classifyFunc: classifyFunc,
		model:        NewPredictiveModel(),
	}, nil
}

// Start begins the transaction monitoring process
func (tms *TransactionMonitoringSystem) Start(ctx context.Context, wg *sync.WaitGroup) {
	defer wg.Done()
	ticker := time.NewTicker(1 * time.Minute)
	defer ticker.Stop()

	for {
		select {
		case <-ticker.C:
			tms.monitorTransactions()
		case <-ctx.Done():
			tms.natsConn.Close()
			close(tms.alertCh)
			return
		}
	}
}

// monitorTransactions fetches and processes recent transactions
func (tms *TransactionMonitoringSystem) monitorTransactions() {
	transactions, err := tms.fetchRecentTransactions()
	if err != nil {
		log.Println("Error fetching transactions:", err)
		return
	}

	var wg sync.WaitGroup
	for _, tx := range transactions {
		wg.Add(1)
		go func(tx Transaction) {
			defer wg.Done()
			tms.processTransaction(tx)
		}(tx)
	}
	wg.Wait()
}

// processTransaction processes a single transaction
func (tms *TransactionMonitoringSystem) processTransaction(tx Transaction) {
	tx.Category = tms.classifyFunc(tx)
	if tms.isAnomalous(tx) {
		anomaly := Anomaly{
			TransactionID: tx.ID,
			DetectedAt:    time.Now(),
			Reason:        "Anomalous transaction detected",
		}
		tms.alertCh <- anomaly
	}

	if err := tms.updateTransactionCategory(tx); err != nil {
		log.Println("Error updating transaction category:", err)
	}
}

// fetchRecentTransactions retrieves recent transactions from the database
func (tms *TransactionMonitoringSystem) fetchRecentTransactions() ([]Transaction, error) {
	rows, err := tms.db.Query(`
		SELECT id, user_id, timestamp, amount, type, status, category 
		FROM transactions 
		WHERE timestamp > NOW() - INTERVAL '1 MINUTE'`)
	if err != nil {
		return nil, err
	}
	defer rows.Close()

	var transactions []Transaction
	for rows.Next() {
		var tx Transaction
		if err := rows.Scan(&tx.ID, &tx.UserID, &tx.Timestamp, &tx.Amount, &tx.Type, &tx.Status, &tx.Category); err != nil {
			return nil, err
		}
		transactions = append(transactions, tx)
	}
	return transactions, rows.Err()
}

// updateTransactionCategory updates the category of a transaction in the database
func (tms *TransactionMonitoringSystem) updateTransactionCategory(tx Transaction) error {
	_, err := tms.db.Exec(`
		UPDATE transactions 
		SET category = $1 
		WHERE id = $2`,
		tx.Category, tx.ID)
	return err
}

// isAnomalous determines if a transaction is anomalous based on predefined criteria
func (tms *TransactionMonitoringSystem) isAnomalous(tx Transaction) bool {
	// Example criteria for anomaly detection (this can be extended with more sophisticated checks)
	if tx.Amount > 10000 { // Large transactions
		return true
	}
	// Add more rules here (e.g., frequency of transactions, unusual transaction types, etc.)
	return false
}

// ServeDashboard serves the dashboard data via an HTTP endpoint
func (tms *TransactionMonitoringSystem) ServeDashboard(w http.ResponseWriter, r *http.Request) {
	tms.dataLock.Lock()
	defer tms.dataLock.Unlock()

	data, err := json.Marshal(tms.dashboardData)
	if err != nil {
		http.Error(w, "Error marshalling dashboard data", http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.Write(data)
}

// updateDashboardData updates the data displayed on the dashboard
func (tms *TransactionMonitoringSystem) updateDashboardData() {
	tms.dataLock.Lock()
	defer tms.dataLock.Unlock()

	totalTransactions, err := tms.getTotalTransactions()
	if err != nil {
		log.Println("Error getting total transactions:", err)
		return
	}

	anomalies, err := tms.getRecentAnomalies()
	if err != nil {
		log.Println("Error getting recent anomalies:", err)
		return
	}

	recentTransactions, err := tms.getRecentTransactions()
	if err != nil {
		log.Println("Error getting recent transactions:", err)
		return
	}

	tms.dashboardData = DashboardData{
		TotalTransactions: totalTransactions,
		Anomalies:         anomalies,
		RecentTransactions: recentTransactions,
	}
}

// getTotalTransactions retrieves the total number of transactions from the database
func (tms *TransactionMonitoringSystem) getTotalTransactions() (int, error) {
	var total int
	err := tms.db.QueryRow(`
		SELECT COUNT(*) 
		FROM transactions`).Scan(&total)
	return total, err
}

// getRecentAnomalies retrieves recent anomalies from the database
func (tms *TransactionMonitoringSystem) getRecentAnomalies() ([]Anomaly, error) {
	rows, err := tms.db.Query(`
		SELECT transaction_id, detected_at, reason 
		FROM anomalies 
		WHERE detected_at > NOW() - INTERVAL '1 DAY'`)
	if err != nil {
		return nil, err
	}
	defer rows.Close()

	var anomalies []Anomaly
	for rows.Next() {
		var an Anomaly
		if err := rows.Scan(&an.TransactionID, &an.DetectedAt, &an.Reason); err != nil {
			return nil, err
		}
		anomalies = append(anomalies, an)
	}
	return anomalies, rows.Err()
}

// getRecentTransactions retrieves recent transactions from the database
func (tms *TransactionMonitoringSystem) getRecentTransactions() ([]Transaction, error) {
	rows, err := tms.db.Query(`
		SELECT id, user_id, timestamp, amount, type, status, category 
		FROM transactions 
		WHERE timestamp > NOW() - INTERVAL '1 DAY'`)
	if err != nil {
		return nil, err
	}
	defer rows.Close()

	var transactions []Transaction
	for rows.Next() {
		var tx Transaction
		if err := rows.Scan(&tx.ID, &tx.UserID, &tx.Timestamp, &tx.Amount, &tx.Type, &tx.Status, &tx.Category); err != nil {
			return nil, err
		}
		transactions = append(transactions, tx)
	}
	return transactions, rows.Err()
}

// PredictiveModel represents the predictive model for transaction monitoring
type PredictiveModel struct {
	// Add model fields here
}

// NewPredictiveModel initializes a new predictive model
func NewPredictiveModel() *PredictiveModel {
	return &PredictiveModel{
		// Initialize model fields here
	}
}

// Predict predicts if a transaction is suspicious
func (pm *PredictiveModel) Predict(tx Transaction) bool {
	// Add prediction logic here
	return false
}

// Utility functions for secure communication, encryption, and decryption
func generateSalt() ([]byte, error) {
	salt := make([]byte, 16)
	_, err := rand.Read(salt)
	return salt, err
}

func hashPassword(password string, salt []byte) []byte {
	return argon2.IDKey([]byte(password), salt, 1, 64*1024, 4, 32)
}

func encrypt(data, passphrase []byte) ([]byte, error) {
	// Use AES for encryption
	block, err := aes.NewCipher(hashPassword(string(passphrase), nil))
	if err != nil {
		return nil, err
	}
	gcm, err := cipher.NewGCM(block)
	if err != nil {
		return nil, err
	}
	nonce := make([]byte, gcm.NonceSize())
	if _, err = io.ReadFull(rand.Reader, nonce); err != nil {
		return nil, err
	}
	return gcm.Seal(nonce, nonce, data, nil), nil
}

func decrypt(encryptedData, passphrase []byte) ([]byte, error) {
	// Use AES for decryption
	block, err := aes.NewCipher(hashPassword(string(passphrase), nil))
	if err != nil {
		return nil, err
	}
	gcm, err := cipher.NewGCM(block)
	if err != nil {
		return nil, err
	}
	nonceSize := gcm.NonceSize()
	if len(encryptedData) < nonceSize {
		return nil, errors.New("ciphertext too short")
	}
	nonce, ciphertext := encryptedData[:nonceSize], encryptedData[nonceSize:]
	return gcm.Open(nil, nonce, ciphertext, nil)
}

// Ensure secure communication between services
func secureCommunication() {
	// Implement secure communication logic here
}
